handle pre-7 and 7 in FindOpticksOptiX.cmake, more GParts accessors needed for CSGOptiXGGeo/Converter

+++ examples/UseBCM/go.sh
@@ -1,29 +1,16 @@
 #!/bin/bash -l
-##
-## Copyright (c) 2019 Opticks Team. All Rights Reserved.
-##
-## This file is part of Opticks
-## (see https://bitbucket.org/simoncblyth/opticks).
-##
-## Licensed under the Apache License, Version 2.0 (the "License"); 
-## you may not use this file except in compliance with the License.  
-## You may obtain a copy of the License at
-##
-##   http://www.apache.org/licenses/LICENSE-2.0
-##
-## Unless required by applicable law or agreed to in writing, software 
-## distributed under the License is distributed on an "AS IS" BASIS, 
-## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  
-## See the License for the specific language governing permissions and 
-## limitations under the License.
-##
 
+if [ -z "$OPTICKS_PREFIX" -o ! -d "$OPTICKS_PREFIX" ]; then
+    echo $BASH_SOURCE : required envvar OPTICKS_PREFIX  $OPTICKS_PREFIX is not defined or not pointing to an existing directory  
+    exit 1
+fi 
+echo OPTICKS_PREFIX : $OPTICKS_PREFIX
 
-opticks-
 
 sdir=$(pwd)
 name=$(basename $sdir) 
-bdir=/tmp/$USER/opticks/$name/build 
+PREFIX=/tmp/$USER/opticks/$name
+bdir=$PREFIX/build 
 
 rm   -rf $bdir
 mkdir -p $bdir 
@@ -32,12 +19,16 @@ pwd
 
 cmake $sdir \
     -DCMAKE_BUILD_TYPE=Debug \
-    -DCMAKE_PREFIX_PATH=$(opticks-prefix)/externals \
-    -DCMAKE_INSTALL_PREFIX=$(opticks-prefix) \
-    -DCMAKE_MODULE_PATH=$(opticks-home)/cmake/Modules
+    -DCMAKE_PREFIX_PATH=${OPTICKS_PREFIX}/externals \
+    -DCMAKE_INSTALL_PREFIX=$PREFIX 
+
+
+#    -DCMAKE_MODULE_PATH=$(opticks-home)/cmake/Modules
 
 #make
 #make install   
 
 
+
+exit 0
